requirements.txt
google-cloud-bigquery==3.11.4
psycopg2-binary==2.9.1
python-dotenv==1.0.0
google-cloud-storage==2.10.0
flask>=2.0.0
werkzeug==2.3.7
gunicorn==20.1.0
google-cloud-alloydb-connector[pg8000]>=1.7.0
sqlalchemy>=2.0.0
pg8000>=1.29.0
